+++ public/test_server_8_load_balancer.php
@@ -1,0 +1,3 @@
+<?php
+
+print(json_encode($_SERVER, JSON_PRETTY_PRINT));