+++ public/test_server_8_load_redirect.php
@@ -1,0 +1,4 @@
+<?php
+
+header('Location: https://webapp.dev/test_server_8_load_balancer.php');
+exit;