+++ tests/Services/Nginx/HttpTest.php
@@ -2,7 +2,9 @@
 
 class HttpTest extends TestCase
 {
-
+    /**
+    *@group prod
+    */
     public function testHttpRedirectToHttps()
     {
         $response= $this->getFullResponseFromURL('http://nginxhttp/');
@@ -10,8 +12,9 @@ class HttpTest extends TestCase
         $this->assertContains('HTTP/1.1 301 Moved Permanently', $response);
         $this->assertContains('Location: https://nginxhttp/', $response);
     }
-
     /**
+    *@group prod
+    *
     *Make sure the acme-challenge directory is served over http and requester is not redirected to https.
     */
     public function testServingLetsencryptAcmeChallenge()
@@ -21,6 +24,9 @@ class HttpTest extends TestCase
         $this->assertContains('HTTP/1.1 404 Not Found', $response);
     }
 
+    /**
+    *@group prod
+    */
     public function testDontSendNginxVersionNumberInHeader()
     {
         $response= $this->getFullResponseFromURL('http://nginxhttp/');